NOTE!!!!!!!!!!!!!!!!! !!!!!!!!!!!!!!!!!!!!!
Great results for extraction and qlassification with local LLMs.
Next iteration include cloud providers, from OpenRouter.

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   aileana_helper_SQL_queries.sql
	modified:   helper_llm_main.py

Changes not staged for commit:
	modified:   app.py
	modified:   data/prompts/examples_job_industry.json
	deleted:    error_log.txt
	modified:   helper_llm_main.py